Fixing get account and user store
