+++ src/main/webapp/app/components/snackbar/snackbar.tsx
@@ -1,18 +1,15 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import Button from '@material-ui/core/Button';
+import amber from '@material-ui/core/colors/amber';
+import green from '@material-ui/core/colors/green';
+import IconButton from '@material-ui/core/IconButton';
+import SnackbarContent from '@material-ui/core/SnackbarContent';
+import { withStyles } from '@material-ui/core/styles';
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
+import CloseIcon from '@material-ui/icons/Close';
 import ErrorIcon from '@material-ui/icons/Error';
 import InfoIcon from '@material-ui/icons/Info';
-import CloseIcon from '@material-ui/icons/Close';
-import green from '@material-ui/core/colors/green';
-import amber from '@material-ui/core/colors/amber';
-import IconButton from '@material-ui/core/IconButton';
-import Snackbar from '@material-ui/core/Snackbar';
-import SnackbarContent from '@material-ui/core/SnackbarContent';
 import WarningIcon from '@material-ui/icons/Warning';
-import { withStyles } from '@material-ui/core/styles';
+import classNames from 'classnames';
+import React from 'react';
 
 const variantIcon = {
   success: CheckCircleIcon,
@@ -81,77 +78,4 @@ class MySnackbarContent extends React.Component<IMySnackbarContentProps> {
   }
 }
 
-const MySnackbarContentWrapper = withStyles(styles1)(MySnackbarContent);
-
-const styles2 = theme => ({
-  margin: {
-    margin: theme.spacing.unit
-  }
-});
-
-interface ICustomizedSnackbarsProps {
-  classes: any;
-}
-
-class CustomizedSnackbars extends React.Component<ICustomizedSnackbarsProps> {
-  state = {
-    open: false
-  };
-
-  handleClick = () => {
-    this.setState({ open: true });
-  };
-
-  handleClose = (event, reason) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-
-    this.setState({ open: false });
-  };
-
-  render() {
-    const { classes } = this.props;
-
-    return (
-      <div>
-        <Button className={classes.margin} onClick={this.handleClick}>
-          Open success snackbar
-        </Button>
-        <Snackbar
-          anchorOrigin={{
-            vertical: 'bottom',
-            horizontal: 'center'
-          }}
-          open={this.state.open}
-          autoHideDuration={3000}
-          onClose={this.handleClose}
-        >
-          <MySnackbarContentWrapper variant="error" className={classes.margin} message="This is an error message!" />
-        </Snackbar>
-        {/* <MySnackbarContentWrapper
-          variant="error"
-          className={classes.margin}
-          message="This is an error message!"
-        />
-        <MySnackbarContentWrapper
-          variant="warning"
-          className={classes.margin}
-          message="This is a warning message!"
-        />
-        <MySnackbarContentWrapper
-          variant="info"
-          className={classes.margin}
-          message="This is an information message!"
-        />
-        <MySnackbarContentWrapper
-          variant="success"
-          className={classes.margin}
-          message="This is a success message!"
-        /> */}
-      </div>
-    );
-  }
-}
-
-export default withStyles(styles2)(CustomizedSnackbars);
+export const MySnackbarContentWrapper = withStyles(styles1)(MySnackbarContent);
