+++ src/main/webapp/app/stores/user-store.tsx
@@ -1,22 +1,19 @@
-import { computed } from 'mobx';
 import * as UserInterface from 'app/shared/model/user.model';
+import { computed, observable, toJS } from 'mobx';
 
 type IUser = UserInterface.IUser;
 
 class UserStore {
+  @observable
   private innerUser: IUser = {};
 
   @computed
   get user(): IUser {
-    return this.innerUser;
+    return toJS(this.innerUser);
   }
 
   set user(user: IUser) {
     this.innerUser = user;
-  }
-
-  get() {
-    return this.innerUser;
   }
 }
