+++ src/main/webapp/app/views/home/home.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
-import { observer } from 'mobx-react';
-import { TextField, Button } from '@material-ui/core';
-import './login.scss';
+import { Button, TextField } from '@material-ui/core';
 import { apiUtil } from 'app/api/api';
 import { userStore } from 'app/stores/user-store';
+import { observer } from 'mobx-react';
+import React from 'react';
+import './login.scss';
 
 @observer
 export class VibeHome extends React.Component {
